Move download_content_pack to i18n.base.
Will be used by the create_dummy_language_pack command.
